Make #amqqueue{} a private record
The #amqqueue{} record is now isolated in the `amqqueue` module: all
accesses and modifications to the record are made through this helper.

The type `rabbit_types:amqqueue()` is now `amqqueue:amqqueue()`.

`amqqueue` knows about the new #amqqueue{} record introduced with quorum
queues. If the actual record is the old one (as defined in RabbitMQ
3.7.x), it calls `amqqueue_v1` which knows about the old definition.

When it needs to produce an #amqqueue{} record, it verifies if it is
allowed to, using the new Feature flags subsystem and the `quorum_queue`
feature flag specifically, and proceeds. If an old format #amqqueue{} is
required, the creation is deferred to `amqqueue_v1`.

The new modules come with a couple headers: they provide macros to
replace pattern matching with `when` conditions. The `amqqueue_v*.hrl`
headers are generated using the `macros()` functions in `amqqueue` and
`amqqueue_v1` modules.

Some operations now depend on the state of the `quorum_queue` feature
flag. In particular, creating a quorum queue is denied if the feature
flag is disabled.

In the process, the following modules were moved from rabbitmq-common to
rabbitmq-server:
  * rabbit_backing_queue
  * rabbit_queue_master_locator

Likewise, the few functions were copied from `rabbit_misc` to
`rabbit_amqqueue` because they depend on `amqqueue`. They don't really
need to live in rabbitmq-common and they prevent further progress with
the feature flags: feature flags need informations about the cluster.

[#159298729]